Before refactoring
